Set machine name

+++ install.sh
@@ -5,8 +5,23 @@
 # @author Charlie Revett
 
 echo "> starting install.sh"
-echo "> creating symlinks"
+echo "> setting machine name"
 
+MACHINE_NAME_FILE=~/projects/github.com/revett/dotfiles/machine_name.txt
+if [ ! -f "$MACHINE_NAME_FILE" ]; then
+  echo "$MACHINE_NAME_FILE does not exist."
+  exit 1
+fi
+
+MACHINE_NAME=$(cat $MACHINE_NAME_FILE)
+echo "> using \"$MACHINE_NAME\" as machine name"
+
+sudo scutil --set HostName $MACHINE_NAME
+sudo scutil --set LocalHostName $MACHINE_NAME
+sudo scutil --set ComputerName $MACHINE_NAME
+dscacheutil -flushcache
+
+echo "> creating symlinks"
 # Shell
 ln -sf ~/projects/github.com/revett/dotfiles/.zshrc ~/.zshrc
